refactor: profile screen business logic moved to hooks
